changing size_t to int to remove static_cast<size_t> to optimize more

+++ BFS.h
@@ -5,10 +5,10 @@
 #include <vector>
 #include <queue>
 
-void BFS(std::vector<size_t> &init, const std::vector<size_t> &goal, const int &depth_limit = -1);
-bool is_goal(const std::vector<size_t> &state, const std::vector<size_t> &goal);
-void solved(size_t steps, const std::vector<size_t> &goal);
-void print_puzzle(size_t steps, const std::vector<size_t> &state);
-int getInvCount(const std::vector<size_t> &arr);
-bool isSolvable(const std::vector<size_t> &arr, const std::vector<size_t> &goal);
+void BFS(std::vector<int> &init, const std::vector<int> &goal, const int &depth_limit = -1);
+bool is_goal(const std::vector<int> &state, const std::vector<int> &goal);
+void solved(int steps, const std::vector<int> &goal);
+void print_puzzle(int steps, const std::vector<int> &state);
+int getInvCount(const std::vector<int> &arr);
+bool isSolvable(const std::vector<int> &arr, const std::vector<int> &goal);
 #endif
