+++ DFS.h
@@ -5,6 +5,6 @@
 #include <vector>
 #include <stack>
 
-void DFS(std::vector<size_t> &init, const std::vector<size_t> &goal, const int &depth_limit = -1);
+void DFS(std::vector<int> &init, const std::vector<int> &goal, const int &depth_limit = -1);
 
 #endif
