+++ another_way_BFS.h
@@ -5,10 +5,10 @@
 #include <vector>
 #include <queue>
 
-void another_algorithm_BFS(const std::vector<size_t> &init, const std::vector<size_t> &goal);
-bool is_goal_1(const std::vector<size_t> &state, const std::vector<size_t> &goal);
-bool is_goal_2(const std::vector<size_t> &state, const std::vector<size_t> &goal);
-bool is_goal_3(const std::vector<size_t> &state, const std::vector<size_t> &goal);
-bool is_goal_4(const std::vector<size_t> &state, const std::vector<size_t> &goal);
+void another_algorithm_BFS(const std::vector<int> &init, const std::vector<int> &goal);
+bool is_goal_1(const std::vector<int> &state, const std::vector<int> &goal);
+bool is_goal_2(const std::vector<int> &state, const std::vector<int> &goal);
+bool is_goal_3(const std::vector<int> &state, const std::vector<int> &goal);
+bool is_goal_4(const std::vector<int> &state, const std::vector<int> &goal);
 
 #endif
